Autoconfiguration, pat rewrite : we check the transport_stream id. This should solve problems with shared pids
When a PMT is common to multiple channels, autoconfiguration can fail finding the right pids. This is corrected by
checking the transport stream id. A new option option 'ts_id' is created when autoconfiguration full is not used.

The pat rewrite also take care of the transport stream id to announce only this program even if the PMT is shared