Make now_time non-global.
There's no good reason to have non_time global, except to create a huge amount
of races; and worse, uint64_t is a non-atomic type on 32-bit x86, causing
potentially invalid values to be read as the time. Simply make it local to
each function that needs it, since they were updating it for every read anyway.

Signed-off-by: Brice DUBOST <mumudvb@braice.net>