Introduce a temporary variable "channel" in some cases.
Some places, some long expression such as params->chan_and_pids->channels[curr_channel]
is used over and over again; simply add a temporary variable holding the pointer to the
channel in those cases. This removes a lot of unneccessary clutter that's in the way
when trying to understand the code.

Signed-off-by: Brice DUBOST <mumudvb@braice.net>