Fix UTF-8 output in state.xml.
First of all, do not clear non-ASCII characters from the output.

Second; BOMs are not recommended for UTF-8 (the only reason they're
in use at all, is that Microsoft for some reason decided to
start using them as some sort of UTF-8 marker without talking to the Unicode
Consortium), so do not include them.

Signed-off-by: Brice DUBOST <mumudvb@braice.net>