Make the Interrupted variable thread-safe.
Interrupted is set from various places as a generic error variable
any form of locking; sometimes, in contexts that can set it to 0 even though it
was something else already.  This can cause errors to be masked, and trips up
Helgrind a lot.

Make sure to always set it atomically, and only if it is zero (so previous
errors are not overwritten; we always get the first one). Also, never set it
to zero no matter what.

Signed-off-by: Brice DUBOST <mumudvb@braice.net>