Rework ring buffer locking.
Instead of having one mutex per variable, have one that protects everything
except the CWs (which have their own locks). In particular, this protects
time_decsa and time_send, which were not at all protected before. (This could
lead to the decsa thread hanging when time_decsa[] was changed between the
check against "now_time" and the sleep time calculation, and we thus tried to
sleep a negative amount.)

Also, remove the redundant variable ring_buffer_num_packets, which is only
the sum of to_descramble and to_send anyway.

Signed-off-by: Brice DUBOST <mumudvb@braice.net>