Start SCAM threads on channels, not PIDs.
This code used to check if a SCAM thread was stared for a given PID,
but the threads operate per-channel. (This is wrong, by the way;
there exist channels where different PIDs need different keys,
but I'm not going to fix that right now.) This would frequently make multiple
decsa and send threads being spawned for the same channel, which would stomp on
each other. Change so that we put the check on the channel instead of the PID,
which fixes the problem.

Signed-off-by: Brice DUBOST <mumudvb@braice.net>