replace mumudvb.braice.net by mumudvb.net
mumudvb.braice.net seems to be dead, the new hosting is at mumudvb.net.
Apart from the change to the openwrt packaging file this seems to be the
right thing to do. I changed openwrt/packages/mumudvb/Makefile anyhow
because otherwise this would be the only instance mentioning the old url
and the new isn't worse than the new one.

Closes: #212